requirements.txt
asgiref==3.2.10
Django==3.1.2
image==1.5.33
Pillow==8.0.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
